remove connection manager and refactored messages for DIA-2628-handling-no-internet
