+++ cmplibrary/src/main/java/com/sourcepoint/cmplibrary/creation/Builder.kt
@@ -21,8 +21,6 @@ import com.sourcepoint.cmplibrary.data.local.DataStorage
 import com.sourcepoint.cmplibrary.data.local.DataStorageCcpa
 import com.sourcepoint.cmplibrary.data.local.DataStorageGdpr
 import com.sourcepoint.cmplibrary.data.local.create
-import com.sourcepoint.cmplibrary.data.network.connection.ConnectionManager
-import com.sourcepoint.cmplibrary.data.network.connection.create
 import com.sourcepoint.cmplibrary.data.network.converter.JsonConverter
 import com.sourcepoint.cmplibrary.data.network.converter.create
 import com.sourcepoint.cmplibrary.data.network.converter.genericFail
@@ -79,10 +77,9 @@ class Builder {
         val errorManager = errorMessageManager(campaignManager, client)
         val logger = spc.logger ?: createLogger(errorManager)
         val jsonConverter = JsonConverter.create()
-        val connManager = ConnectionManager.create(appCtx)
         val responseManager = ResponseManager.create(jsonConverter, logger)
         val networkClient = networkClient(appCtx, okHttpClient, responseManager, logger)
-        val viewManager = ViewsManager.create(activityWeakRef, connManager, spc.messageTimeout)
+        val viewManager = ViewsManager.create(activityWeakRef, spc.messageTimeout)
         val execManager = ExecutorManager.create(appCtx)
         val urlManager: HttpUrlManager = HttpUrlManagerSingleton
         val consentManagerUtils: ConsentManagerUtils = ConsentManagerUtilsImpl(campaignManager, dataStorage)
