--- cmplibrary/src/main/java/com/sourcepoint/cmplibrary/data/network/connection/ConnectionManager.kt
@@ -1,6 +1,0 @@
-package com.sourcepoint.cmplibrary.data.network.connection
-
-internal interface ConnectionManager {
-    val isConnected: Boolean
-    companion object
-}
