+++ cmplibrary/src/main/java/com/sourcepoint/cmplibrary/exception/ExceptionDescriptions.kt
@@ -1,0 +1,8 @@
+package com.sourcepoint.cmplibrary.exception
+
+class ExceptionDescriptions {
+
+    companion object {
+        const val EXCEPTION_NO_INTERNET_CONNECTION_DESCRIPTION = "No Internet connection."
+    }
+}
