+++ cmplibrary/src/main/java/com/sourcepoint/cmplibrary/util/ViewsManager.kt
@@ -9,7 +9,6 @@ import com.sourcepoint.cmplibrary.core.Either
 import com.sourcepoint.cmplibrary.core.web.CampaignModel
 import com.sourcepoint.cmplibrary.core.web.ConsentWebView
 import com.sourcepoint.cmplibrary.core.web.IConsentWebView
-import com.sourcepoint.cmplibrary.data.network.connection.ConnectionManager
 import com.sourcepoint.cmplibrary.exception.WebViewCreationException
 import com.sourcepoint.cmplibrary.model.exposed.MessageSubCategory
 import java.lang.ref.WeakReference
@@ -51,17 +50,14 @@ internal interface ViewsManager {
  */
 internal fun ViewsManager.Companion.create(
     actWeakReference: WeakReference<Activity>,
-    connectionManager: ConnectionManager,
     messageTimeout: Long
 ): ViewsManager = ViewsManagerImpl(
     weakReference = actWeakReference,
-    connectionManager = connectionManager,
     messageTimeout = messageTimeout
 )
 
 private class ViewsManagerImpl(
     val weakReference: WeakReference<Activity>,
-    val connectionManager: ConnectionManager,
     val messageTimeout: Long
 ) : ViewsManager {
 
@@ -132,7 +128,6 @@ private class ViewsManagerImpl(
             check {
                 ConsentWebView(
                     context = it,
-                    connectionManager = connectionManager,
                     jsClientLib = jsReceiverDelegate,
                     logger = lib.pLogger,
                     executorManager = lib.executor,
@@ -155,7 +150,6 @@ private class ViewsManagerImpl(
             check {
                 ConsentWebView(
                     context = it,
-                    connectionManager = connectionManager,
                     jsClientLib = jsReceiverDelegate,
                     logger = lib.pLogger,
                     executorManager = lib.executor,
