+++ cmplibrary/src/test/java/com/sourcepoint/cmplibrary/SpConsentLibImplTest.kt
@@ -9,16 +9,15 @@ import com.sourcepoint.cmplibrary.core.Either
 import com.sourcepoint.cmplibrary.core.ExecutorManager
 import com.sourcepoint.cmplibrary.data.Service
 import com.sourcepoint.cmplibrary.data.local.DataStorage
-import com.sourcepoint.cmplibrary.data.network.connection.ConnectionManager
 import com.sourcepoint.cmplibrary.data.network.converter.JsonConverter
 import com.sourcepoint.cmplibrary.data.network.util.HttpUrlManager
+import com.sourcepoint.cmplibrary.data.network.util.isInternetConnected
 import com.sourcepoint.cmplibrary.exception.CampaignType.GDPR
 import com.sourcepoint.cmplibrary.exception.Logger
 import com.sourcepoint.cmplibrary.model.Campaign
 import com.sourcepoint.cmplibrary.model.PMTab
 import com.sourcepoint.cmplibrary.model.exposed.MessageSubCategory.* //ktlint-disable
 import com.sourcepoint.cmplibrary.util.ViewsManager
-import com.sourcepoint.cmplibrary.util.extensions.isInternetConnected
 import io.mockk.MockKAnnotations
 import io.mockk.every
 import io.mockk.impl.annotations.MockK
@@ -41,9 +40,6 @@ class SpConsentLibImplTest {
 
     @MockK
     private lateinit var jsonConverter: JsonConverter
-
-    @MockK
-    private lateinit var connManager: ConnectionManager
 
     @MockK
     private lateinit var dataStorage: DataStorage
